wip: tidying code
fixed remaining @return where @returns is preferred for javadoc style.
removed ideal line edit and replaced with QWidget as placeholder for correct alignment in the latency settings page

+++ src/components/RouteAnalyser/GraphLatencyLayer.h
@@ -70,7 +70,7 @@ namespace Nedrysoft::RouteAnalyser {
             bool m_useGradient;
 
             static QMap<QString, QPixmap> m_buffers;
-            static QMap<QString, int> m_age;
+            static QMap<QString, uint64_t> m_age;
     };
 }
